@@ -6,31 +6,33 @@ import android.content.Intent;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.ItemTouchHelper;
+
 import android.view.View;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
+
+import com.google.android.material.navigation.NavigationView;
+
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import android.view.MenuItem;
 import android.widget.EditText;
-import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import hu.bme.aut.weatherapp.adapter.CitiesAdapter;
 import hu.bme.aut.weatherapp.touch.CityItemTouchHelperCallback;
 import hu.bme.aut.weatherapp.data.City;
 
-public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener
-{
+public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
     public static final String CITY_SELECTED = "CITY_SELECTED";
 
     private RecyclerView recyclerCities;
@@ -39,8 +41,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     private int positionOfCity = -1;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState)
-    {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
@@ -48,31 +49,28 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     }
 
 
-    public void setUpUI()
-    {
+    public void setUpUI() {
         setupRecyclerView();
         setUpToolBar();
     }
 
-    public void setUpToolBar()
-    {
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+    public void setUpToolBar() {
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
                 this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.setDrawerListener(toggle);
+        drawer.addDrawerListener(toggle);
         toggle.syncState();
 
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        NavigationView navigationView = findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
     }
 
 
-    private void setupRecyclerView()
-    {
-        recyclerCities = (RecyclerView) findViewById(R.id.recyclerCities);
+    private void setupRecyclerView() {
+        recyclerCities = findViewById(R.id.recyclerCities);
         recyclerCities.setHasFixedSize(true);
 
         final LinearLayoutManager layoutManager = new LinearLayoutManager(this);
@@ -81,21 +79,17 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         citiesAdapter = new CitiesAdapter(this);
         recyclerCities.setAdapter(citiesAdapter);
 
-        final FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener()
-        {
+        final FloatingActionButton fab = findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View view)
-            {
+            public void onClick(View view) {
                 showAddCityDialog();
             }
         });
 
-        recyclerCities.addOnScrollListener(new RecyclerView.OnScrollListener()
-        {
+        recyclerCities.addOnScrollListener(new RecyclerView.OnScrollListener() {
             @Override
-            public void onScrolled(RecyclerView recyclerView, int dx, int dy)
-            {
+            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
                 super.onScrolled(recyclerView, dx, dy);
                 if (dy > 0 && fab.getVisibility() == View.VISIBLE) {
                     fab.hide();
@@ -104,6 +98,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
                 }
             }
         });
+
         // adding touch support
         ItemTouchHelper.Callback callback = new CityItemTouchHelperCallback(citiesAdapter);
         ItemTouchHelper touchHelper = new ItemTouchHelper(callback);
@@ -111,8 +106,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
     }
 
-    public void showAddCityDialog()
-    {
+    public void showAddCityDialog() {
         final AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle("City Name");
 
@@ -121,8 +115,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
         builder.setPositiveButton("Ok", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialog, int which)
-            {
+            public void onClick(DialogInterface dialog, int which) {
                 citiesAdapter.addCity(new City(etCityText.getText().toString()));
 
                 recyclerCities.scrollToPosition(0);
@@ -141,67 +134,52 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         builder.show();
     }
 
-    public void openWeatherActivity(int index, String city_selected)
-    {
-        if(isOnline(getApplicationContext()))
-        {
+    public void openWeatherActivity(int index, String city_selected) {
+        if (isOnline(getApplicationContext())) {
             positionOfCity = index;
 
             Intent intentShowWeather = new Intent(this, WeatherActivity.class);
-
             intentShowWeather.putExtra(CITY_SELECTED, city_selected);
-
             startActivity(intentShowWeather);
-        }
-        else
-        {
+        } else {
             Toast.makeText(MainActivity.this, "No internet connection", Toast.LENGTH_LONG).show();
         }
 
     }
 
     @Override
-    public void onBackPressed()
-    {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-        if (drawer.isDrawerOpen(GravityCompat.START))
-        {
+    public void onBackPressed() {
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        if (drawer.isDrawerOpen(GravityCompat.START)) {
             drawer.closeDrawer(GravityCompat.START);
-        }
-        else
-        {
+        } else {
             super.onBackPressed();
         }
     }
 
     @SuppressWarnings("StatementWithEmptyBody")
     @Override
-    public boolean onNavigationItemSelected(MenuItem item)
-    {
+    public boolean onNavigationItemSelected(MenuItem item) {
         // Handle navigation view item clicks here.
         int id = item.getItemId();
 
-        if (id == R.id.nav_addcity)
-        {
+        if (id == R.id.nav_addcity) {
             showAddCityDialog();
-        }
-        else if (id == R.id.nav_about)
-        {
+        } else if (id == R.id.nav_about) {
             Toast.makeText(this, "Created by: David Bojkovski", Toast.LENGTH_LONG).show();
         }
 
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
+
         return true;
     }
 
-    public static boolean isOnline(Context context)
-    {
+    public static boolean isOnline(Context context) {
         ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
         NetworkInfo netInfo = cm.getActiveNetworkInfo();
-        if (netInfo != null && netInfo.isConnected())
-        {
+
+        if (netInfo != null && netInfo.isConnected()) {
             return true;
         }
         return false;
