@@ -1,12 +1,12 @@
 package hu.bme.aut.weatherapp.adapter;
 
 import android.content.Context;
-import android.media.Image;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CheckBox;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -19,22 +19,19 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder> implements CityTouchHelperAdapter
-{
+public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder> implements CityTouchHelperAdapter {
     private List<City> citiesList;
 
     private Context context;
 
-    public CitiesAdapter(Context context)
-    {
+    public CitiesAdapter(Context context) {
         this.context = context;
 
-       citiesList = new ArrayList<City>();
+        citiesList = new ArrayList<>();
     }
 
     @Override
-    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType)
-    {
+    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         View rowView = LayoutInflater.from(parent.getContext()).inflate(
                 R.layout.city_row, parent, false);
 
@@ -42,16 +39,13 @@ public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder
     }
 
     @Override
-    public void onBindViewHolder(final ViewHolder holder, int position)
-    {
+    public void onBindViewHolder(final ViewHolder holder, int position) {
         holder.tvCity.setText(citiesList.get(position).getName());
 
-        holder.itemView.setOnClickListener(new View.OnClickListener()
-        {
+        holder.itemView.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View v)
-            {
-                ((MainActivity)context).openWeatherActivity(holder.getAdapterPosition(),
+            public void onClick(View v) {
+                ((MainActivity) context).openWeatherActivity(holder.getAdapterPosition(),
                         citiesList.get(holder.getAdapterPosition()).getName()
                 );
             }
@@ -73,17 +67,11 @@ public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder
     public void onItemDismiss(int position) {
         citiesList.remove(position);
 
-        // refreshes the whole list
-        //notifyDataSetChanged();
-        // refreshes just the relevant part that has been deleted
         notifyItemRemoved(position);
     }
 
     @Override
     public void onItemMove(int fromPosition, int toPosition) {
-        /*todoList.add(toPosition, todoList.get(fromPosition));
-        todoList.remove(fromPosition);*/
-
         if (fromPosition < toPosition) {
             for (int i = fromPosition; i < toPosition; i++) {
                 Collections.swap(citiesList, i, i + 1);
@@ -98,17 +86,15 @@ public class CitiesAdapter extends RecyclerView.Adapter<CitiesAdapter.ViewHolder
         notifyItemMoved(fromPosition, toPosition);
     }
 
-    public static class ViewHolder extends RecyclerView.ViewHolder
-    {
+    public static class ViewHolder extends RecyclerView.ViewHolder {
         private ImageView ivCity;
         private TextView tvCity;
 
-        public ViewHolder(View itemView)
-        {
+        public ViewHolder(View itemView) {
             super(itemView);
 
-            ivCity = (ImageView) itemView.findViewById(R.id.ivCity);
-            tvCity = (TextView) itemView.findViewById(R.id.tvCity);
+            ivCity = itemView.findViewById(R.id.ivCity);
+            tvCity = itemView.findViewById(R.id.tvCity);
         }
     }
 }
