@@ -1,7 +1,7 @@
 package hu.bme.aut.weatherapp.touch;
 
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.ItemTouchHelper;
 
 
 public class CityItemTouchHelperCallback extends ItemTouchHelper.Callback
